requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
django-extensions==3.2.3
python-dotenv==1.0.1